+++ test/shopify-cli/theme/theme_admin_api_test.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 require "test_helper"
 require "shopify_cli/theme/theme_admin_api"
+require "shopify_cli/theme/theme_access_api"
 
 module ShopifyCLI
   module Theme
@@ -31,7 +32,7 @@ module ShopifyCLI
           path: path,
         }
 
-        expect_correct_request_arguments(request_params)
+        expect_admin_api_request_with_params(request_params)
 
         @api_client.get(
           path: path
@@ -50,7 +51,7 @@ module ShopifyCLI
           body: body,
         }
 
-        expect_correct_request_arguments(request_params)
+        expect_admin_api_request_with_params(request_params)
 
         @api_client.put(
           path: path,
@@ -70,7 +71,7 @@ module ShopifyCLI
           body: body,
         }
 
-        expect_correct_request_arguments(request_params)
+        expect_admin_api_request_with_params(request_params)
 
         @api_client.post(
           path: path,
@@ -85,7 +86,7 @@ module ShopifyCLI
           path: path,
         }
 
-        expect_correct_request_arguments(request_params)
+        expect_admin_api_request_with_params(request_params)
 
         @api_client.delete(
           path: path
@@ -124,7 +125,7 @@ module ShopifyCLI
           path: path,
         }
 
-        expect_correct_request_arguments(request_params)
+        expect_admin_api_request_with_params(request_params)
 
         @api_client.post(
           path: path,
@@ -229,9 +230,29 @@ module ShopifyCLI
         refute(@api_client.send(:unauthorized_response?, error))
       end
 
+      def test_theme_access_api_is_used_with_theme_access_password
+        shop = "testshop.myshopify.io"
+        password = "shptka_XXX"
+        Environment.expects(:store).returns(shop)
+        Environment.expects(:admin_auth_token).returns(password)
+        api_client = ThemeAdminAPI.new(@ctx)
+        request_params = {
+          method: "POST",
+          path: "themes.json",
+          query: "query",
+          body: JSON.generate({ theme: {} }),
+          token: password,
+        }
+
+        ThemeAccessAPI.expects(:rest_request)
+          .with(@ctx, shop: shop, api_version: @api_version, **request_params)
+
+        api_client.send(:rest_request, **request_params)
+      end
+
       private
 
-      def expect_correct_request_arguments(request_params)
+      def expect_admin_api_request_with_params(request_params)
         ShopifyCLI::AdminAPI.expects(:rest_request)
           .with(@ctx, shop: @shop, api_version: @api_version, **request_params)
       end